moved files location, will start to-do list now
