+++ app/dashboard/student/StudentToDo.tsx
@@ -1,0 +1,6 @@
+
+export default function StudentTodo(){
+    return (
+        <h1>To-do</h1>
+    )
+}
